+++ lib/SimpleModbusSlaveSoftwareSerial/SimpleModbusSlaveSoftwareSerial.h
@@ -84,6 +84,6 @@
 
 // function definitions
 void modbus_configure(SoftwareSerial* comPort, long baud, unsigned char _slaveID, unsigned char _TxEnablePin, unsigned int _holdingRegsSize);
-unsigned int modbus_update(unsigned int *holdingRegs);
+unsigned int modbus_update(uint16_t *holdingRegs);
 
 #endif
